refactors answers codes to parse user data from jwt token. Moves parseToken to seperate file to be imported to avoid duplication
